% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset_info.R
\name{get_asset_info}
\alias{get_asset_info}
\title{Get asset information}
\usage{
get_asset_info()
}
\value{
list of asset names and their information
}
\description{
Currently only works with defaults; returns all information
about all asset classes
}
