% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ohlc_data.R
\name{get_ohlc_data}
\alias{get_ohlc_data}
\title{Get OHLC data}
\usage{
get_ohlc_data(pair = "XBTEUR", interval = 1, since = 0)
}
\arguments{
\item{pair}{asset pair to get OHLC data for}

\item{interval}{time frame interval minutes (optional):
1 (default), 5, 15, 30, 60, 240, 1440, 10080, 21600}

\item{since}{return committed OHLC data since given id (optional) default: 0}
}
\description{
Returns OHLC data for the specified asset and since
}
