% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_order_book.R
\name{get_order_book}
\alias{get_order_book}
\title{Get order book}
\usage{
get_order_book(pair = "XBTEUR", count = NULL)
}
\arguments{
\item{pair}{asset pair to get market depth for}

\item{count}{maximum number of asks/bids (not currently implemented)}
}
\description{
Returns current orders for bid and ask
}
