% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_spread.R
\name{get_recent_spread}
\alias{get_recent_spread}
\title{Get recent spread data}
\usage{
get_recent_spread(pair = "XBTEUR", since = 0)
}
\arguments{
\item{pair}{asset pair to get market depth for}

\item{since}{return trade data since given id (optional; default: 0)}
}
\description{
Returns recent spread data for a specified asset
}
