% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_trades.R
\name{get_recent_trades}
\alias{get_recent_trades}
\title{gets recent trades from a 'since' date}
\usage{
get_recent_trades(pair, since = 0)
}
\arguments{
\item{pair}{asset pair to get trade data for}

\item{since}{return trade data since given id (optional; default: 0)}
}
\description{
returns recent trades for a specifed asset
}
