% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_server_time.R
\name{get_server_time}
\alias{get_server_time}
\title{gets the server time.}
\usage{
get_server_time()
}
\description{
aid to aaproximating the skew time between server and client.
returns unix timestamp and time in RFC1123 format
}
