% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticker_info.R
\name{get_ticker_info}
\alias{get_ticker_info}
\title{gets recent trades from a 'since' date}
\usage{
get_ticker_info(pair_list = "XBTEUR")
}
\arguments{
\item{pair_list}{comma delimite vector of pairs to get ticker info on}
}
\value{
list of pair names and asset info
}
\description{
returns recent trades for a specifed asset
}
