% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tradable_asset_pair.R
\name{get_tradable_asset_pair}
\alias{get_tradable_asset_pair}
\title{gets recent trades from a 'since' date}
\usage{
get_tradable_asset_pair(info = "info", pair_list = "XBTEUR")
}
\arguments{
\item{info}{info to retrieve (optional):
info: all (default); leverage = leverage info;
fees = fees schedule; margin = margin info}

\item{pair_list}{comma delimited list of asset pairs (default: all)}
}
\description{
returns recent trades for a specifed asset
}
