% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenParsing.R
\name{kraken_annotate_parents_at_ranks}
\alias{kraken_annotate_parents_at_ranks}
\title{Map taxonomy IDs to parents at specified ranks}
\usage{
kraken_annotate_parents_at_ranks(
  kraken_report_df,
  taxonomy,
  ranks = c("S", "G", "F", "K"),
  ancestor = "Ancestor"
)
}
\description{
Given a taxonomy table parsed by \code{\link{kraken_taxonomy_parse}},
compute, for each taxon, the closest ancestor whose rank is in a
user-supplied set of ranks.
}
