% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenParsing.R
\name{kraken_fetch_child_taxids}
\alias{kraken_fetch_child_taxids}
\title{Fetch Child Taxids}
\usage{
kraken_fetch_child_taxids(taxonomy, taxid, inclusive = FALSE)
}
\arguments{
\item{taxonomy}{a taxonomy object produce by \code{\link[=kraken_taxonomy_parse]{kraken_taxonomy_parse()}}}

\item{taxid}{ncbi taxonomic id (integer)}

\item{inclusive}{include the supplied taxid.}

\item{verbose}{print more informative messages}
}
\value{
This function returns the same dataframe from \strong{kraken_report_df} produced by \link{kraken_reports_parse} or link{kraken_report_parse}  but with  a new columns describing inclusive descendancey status to a particular taxid
}
\description{
Identify taxids in your \strong{kraken_report_df} that are either equal to the user specified taxonomy ID or one of its descendants/children
This is useful for telling what species are bacterial / viral / belong to a particular genus etc.
\strong{WARNING: } this function relies on inherent properties of kraken reports - make sure you run this ONLY on unsorted dataframes produced by \link{kraken_reports_parse} or link{kraken_report_parse}
}
