% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenParsing.R
\name{kraken_inspect_parse}
\alias{kraken_inspect_parse}
\title{Parse a kraken inspect file}
\usage{
kraken_inspect_parse(path)
}
\arguments{
\item{path}{Path to file produced by k2 inspect}
}
\value{
a data.table representation of the kraken inspect file.
}
\description{
Parse a kraken inspect file
}
\examples{
path <- system.file("example_data/inspect_pluspf_20210517.txt", package = "krakenR")
kraken_inspect_parse(path)
}
