% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KrakenVisualise.R
\name{kraken_annotate_parents}
\alias{kraken_annotate_parents}
\title{Annotate kraken dataframes with parent taxids}
\usage{
kraken_annotate_parents(kraken_report_df, ancestor = "Ancestor")
}
\arguments{
\item{kraken_report_df}{the kraken dataframe produced by \link{kraken_reports_parse} or link{kraken_report_parse}.
Can be filtered to include only ranks of interest (e.g. just genus and species) but must not be sorted in any way. (relies on default sample order).}
}
\value{
a kraken data.frame with ParentTaxonomyID and ParentScientificName columns added
}
\description{
Annotate kraken dataframes with parent taxids
}
\examples{
# Read in all kraken reports from a directory into a data.frame
kreport_dir <- system.file(package="krakenR", "simulated_data/simulated_kraken_reports_inc_zero_counts/")
df_kreports <- kraken_reports_parse(kreport_dir)

# Annotate parent taxids (useful for sunburst visualisation)
df_kreports <- kraken_annotate_parents(df_kreports)

}
